+++ GRDB/Foundation/RowConvertible+Foundation.swift
@@ -1,0 +1,14 @@
+import Foundation
+
+extension RowConvertible {
+    
+    /**
+    NSDictionary initializer.
+    
+    - parameter dictionary: An NSDictionary.
+    */
+    public init(dictionary: NSDictionary) {
+        let row = Row(dictionary: dictionary)
+        self.init(row: row)
+    }
+}